[CIP-545] Exclude admin users from the limit
Admin users created via Google Auth are excluded from the limit.

Unfortunately, the current setup has no clean way of exluding admin users
from the user's count. The users in `ci_tenant_*.user` has no way of being
marked as admin. The only way to know if a user is admin is to check if the
email exists in `ci.admin_user`, which is in another DB table.

For now, this is an accepted limitation, to be addressed in the future.